[FIX] runbot_merge: make staging serialization more useful
Having the staged and merged heads via an easy to reach endpoint is
nice, but having *just* the hashes is not as it requires basically
exhaustively checking commits against repos to see which is
where. That's annoying.

Make the head serializations a map of repository to commit
hash. That's a lot more convenient.