[IMP] runbot_merge: staging form
Link stagings to the staging commits instead of the commits directly
in order to have the repository name and not need to try the commit on
every repository.

Also make the trees editable, in combination with `readonly="1"` this
does not make the trees editable but does prevent clicking on a row to
open a dialog with the linked objects, which is convenient if one
wants to copy the commit hash.