[ADD] forwardport: email reminders
After 6 months of sitting unmerged, forward ports will now trigger a
monthly email.

Or would, if an email server was configured on the mergebot, but
adding this now should allow debugging its working and tracking the
generation of emails through the "Delivery Failed" collection. This
way if we ever decide to actually do the thing we can just enable an
SMTP server.

Fixes #801