[FIX] forwardport: duplicate batch completion
In the case where multiple batch completions are triggered (see 3
commits earlier) one of the issues is that the "already ported" check
is incorrect, as it checks the (source, target) pair against the PR
used to trigger the completion check. That PR *should* be the one PR
which was just added to a batch, but if one of its descendants is used
to re-trigger a port then the dupe check fails and we try to port the
thing again.

Remember to deref' the subject's source in order to perform the dupe
check, as well as set the source for the forwardport we create.

Fixes #1052