[IMP] runbot: introduce matrix for upgrade
The initial way to configure upgrades was working fine and automatically
but does not alway match the reality of what should be supported.

Saying, upgrade from the last main version and last intermediate
(for 18, from 17.0 and 17.4) we may also want to test from 17.2 because
for some reason there are many people in 17.2 taht will need to upgrade
in 18.0. But the transition 17.2->17.3 doesn't make sense anymore
after a while since people will go in 18 imediatly at some point.

This matrix can be generated automatically from the parameters and new
versions should appear automatically, but it is possible to tweak the
values on demand.

[IMP] runbot: limit trigger to same category

Since only reference_batch from the same category are stored, it makes
senses to limit the complement trigger to the same category.

[IMP] runbot: use upgrade matrix

The current way to define the transition to use is made both using the
version (upgrade from, upgrade too) combined with a selection decision
on the config step (Last intermediate, all intermediate, previous major)

This doesn't allow to easily define in a major version that we
should upgrade from some specific saas version but not to specific saas

17.0 -> 18.0
17.3 -> 18.0
17.4 -> 18.0
but not 17.3->17.4

For now by default using the "last intermediate" and "previous major"
we will test

17.0->18.0
17.4->18.0
17.3->17.4
17.0->17.3
17.0->17.4

We an easily disable 'upgrade to' on 17.3 and 17.4 if this is not
support by the plaform anymore

17.0->18.0
17.4->18.0
17.0->17.4

The issue there is that the upgrade from 17.3 to 18.0 was partially
tested through 17.4, but upgrade from 17.4 is not tested at all anymore.

It is also difficult to define coherent version for all upgrades,
regarding upgrade stables (pushing in master in upgrade repo) and
upgrade to current + templates. The upgrade stable only test major
version to reduce load (7.0->18.0) but does not test 17.4->18.0 making
it possible to merge a pr that breaks the upgrade tests in stable.

This pr proposes to change that using upgrade matrixs, allowing to
define precisely the upgrade to test.

The upgrades will still be defined automatically when merging a new
branch, but can be disabled, reenabled, ..