[IMP] runbot: inherit pr target custom triggers
When teams want to work by merging many branches into a singular dev
branch before targetting master, they often require custom triggers
which we have to copy by hand.
With this change we hope to be able to set the custom trigger on the
target branch instead and inherit the custom triggers on the children
branches automatically.