[FIX] runbot_merge: avoid disappearing closed batches from stagings
This is only a temporary / partial fix (see #1100 for further issues)
but it solves the most obvious problem of a batch having a single PR,
being staged, the staging fails, the PR gets closed, and the batch
disappearing from the staging.

The issue remains that stagings don't actually reflect the state as of
the staging, since they link to the updated version of batches rather
than the version as of staging. That probably can't be fixed if we
want to keep track of everything *but* for documentary and display
purposes the staging should store the state of batches and
PRs (e.g. as JSON) as of its creation.

Fixes #1079