[IMP] runbot: add automatic merge mecanism
The current qualifiers are great to find related errors but an
automation is missing. In some case we could says that if we have the
same error method, error message and tour step/ specific qunit, we could
decide to merge them automatically. This is mainly useful to match
cross version errors.

This pr proposes a new model to define a set of rules qualifiers to
match to be able to merge some errors automaticaly.

Note that a solution to make the fingerprint the same was a possible
but this would hide different messages and maybe make the understanding
and debuging more difficult. We want to have two error contents, linked
to the same error.

A set of qualifier can either be set as automerge or not, if the
criteria are unsure.