[FIX] *: incorrect table display when forwardport stuck
An issue which as occurred a few times is that a forward port gets
stuck somehow and for some reason instead of saying that there are
forward ports left to perform the genealogy tables on the PR
dashboards (both image on github and page on the mergebot) ends at the
last merged PR, despite there being no limit anywhere, which is very
confusing.

The reason is that for some reason (maybe a completely unnecessary
optimisation from assuming the fwport would work in all situations)
the dashboard selector slices off the target branches if all the PRs
in the last batch of the genealogy are closed (which makes sense,
there's not going to be further forward ports probably) *or if
they're all merged*.

So if a forward port gets stuck and all the preceding PRs get merged,
the table shows up as "complete" rather than having outstanding /
missing forward ports.

Just ignore the merged case, and let them fall out into the final case
which cuts the sequence off at the limit.