[IMP] forwardport: don't update local branch when forward porting
Concurent ref updates (in the UI, or different crons) can cause issues
and conflicts ("cannot lock ref" errors), and lead to inconsistent
behaviour to boot as ref resolution is non-atomic.

For cherrypicking, this just requires the usual
fetch-branch-with-commits, then referencing the commit branch head we
resolved instead of the branch by name.

Also resolve the root's target manually instead of implicitly, as we
never updated it at all (not even at the top of the method) so it
entirely relied on somebody else doing branch updates to be up to
date, which is... not good. It's not the worst issue in the world as
it just leads to incorrect modify/delete notifications at worst but
still...

Part of #1065