[REM] runbot_merge: maintenance cron
`--aggressive` is unsuitable for the Odoo repo as it runs out of
memory on the server, and `--prune=now` was a huge mistake.

Rather than fix this, remove it, and try to configure the server to do
things right (kinda, we'll need to remember to add setup for new
repos).