[FIX] forwardport: update entire batch if one PR is modified
If a forward port sequence has multiple PRs and one of the PRs needs
to be modified, that PR will be forward-ported to the current version
of each subsequent branch.

If the sequence has been sitting a while, on the RD project odds are
almost 100% that the bases will now be inconsistent, as the sibling
PRs will be based on too old versions of their respective branches.

To mitigate that, re-port the entire batch at every step.

Notable issues: to avoid losing explicit updates, we stop as soon as a
parent chain gets broken across the entire batch.