[IMP] runbot_merge: snapshot batches when staging
Because stagings link to live batches, there can be strange
interactions when prs get added to or removed from batches between
stagings: the staging will always show the "current" state of the
batch, rather than the state at the time of staging.

By snapshotting the list of PRs in each batch as of staging, we should
make the record more reliable (e.g. for later investigations) as
well as avoid the issue of disappearing batches (closed PRs are
considered inactive, which leads to them not being displayed by
default, but we don't want to display PRs which got closed before
staging happened).

Storing the *entire* staging state doesn't seem necessary, so store
just the list of PRs per batch.

Fixes #1100