[IMP] runbot: log docker file identifier
The current log of the docker file used in a branch is only the docker
tag, it is difficult to know what was the exact version used in this
build.

This commit changes that to log a link to the ceurrent identifier in use
as well as adding some pages to consult docker result from frontend.