[FIX] forwardport: correctly count commits on followup update
When updating a child PR (from a push on one of its ancestors), the
forwardbot would count the number of commits in the updated PR using
the target branch as base. However since #1065 there is no telling
when the remote-tracking branch is updated (possibly never anymore).

This causes the forwardport update to miscount, and basically miss
every squash (single-commit PR), which is not amazing even if !squash
is not blocking anymore. Fix by resolving the target branch's HEAD
before counting the number of commits. This can not use
`remote_head` (`ls-remote`) because there is a (pretty wide all things
considered) window during which the target branch can be updated,
which might lead to the `target_head` being a later commit than the
one `child was rebased on, which would then trigger an "invalid
revision range" if we tried to use it without having fetched it.

Having a later commit as lower bound is not an issue, as in a
`gitrevisions(7)` context `..` will give all commits reachable from
the second commit which are not reachable by the first, so implicitly
resolves the merge-base.

Signaled by lse on odoo/odoo#203992: despite having a single commit
the PR was not squash, this was nbot an issue of a missed sync, rather
the chatter log shows that the update from
87ed09d6b07449587ec6e00104f6749304833588 to
1e2809a5ccfe92c63a2a2d6ec436be6e3c09c22a removed the `squash` flag,
and the logs show that the `sync` event from updating the PR was
ignored (because the head had already been updated).