[IMP] *: retrieve user conf from users json
Previously the users configuration for the mergebot tests was embedded
in pytest.ini files, this is problematic for two reasons:

- When testing against github actual, the users conf needs to have
  real identities (and tokens), which precludes a truly shared
  pytest.ini as (as far as I can tell) pytest does not support any
  sort of conf file inheritance.
- When testing against dummy-central, the pytest.ini content needs to
  be duplicated (in a modified form) in order to give DC the accounts
  setup it needs to expose.

These two components make collocating the accounts configuration in
pytest.ini a poor choice from both a safety and developer experience
perspective.

Move the test suite to reuse the users json instead. This way the
configuration to use with DC can be shared with DC, and there can be a
separate and independent users file for interacting with github actual.

Also rewrite the readme to explain why the mergebot makes sense and
document the testing setup.