[IMP] forwardport: split out PR closing in test
Not *entirely* sure why, but apparently in this test at least creating
a PR and immediately closing it reliably causes github (actual) to
send the `close` event before the PR. As a result the PR is created
closed, which doesn't trigger the branch deletion job (maybe it should
but it's probably not a common case so w/e) and the test breaks
because `cbranch` is never deleted.

Splitting it out in two write blocks mitigates the issue.