[FIX] runbot_merge: error message in patching test
Because that's a pain in the ass dummy central doesn't implement the
"smart" git-upload-pack protocol. This turns out to influence the
error message when a fetch fails to find a commit as a smart server
errors with

    fatal: remote error: upload-pack: not our ref {oid}

while on a dumb server, the client is the one which generates an error

    error: Unable to finds {oid} under {repo}
    Cannot obtain needed object {oid}
    error: fetch failed

Stop checking the error message returned by git, so the test can work
against both DC and github actual.