[IMP] runbot: reduce main page size
The main page will display the latest dev branches witch is not useful
for most users. This is also a cause of excessive load on the server
when the main page is crawled (many reachable links).

This commit removes the dev branches from the main page as well as the
other bundle containing a branch.
This commit also adds an option to see all bundle to restore the
previous behaviour. While on it, fixing the config menu:
- add a expires on the server side setted cookies to avoid loosing the
configuration after closing the browser.
- fixes the triggers_ids selection to avoid displaying all of them
when going into more models
- remove the trigger id cookie if the selected ones are the default ones
- allow to archive a category to hide it in the interface when not in
use.