[ADD] runbot_merge: try to merge batches on retarget
This commit restores the old behaviour of switching batch on target
change, in the specific case where there already is a batch for that
label for the new target.

This is intended for the specific use case where a dev makes a
mistakes and creates two PRs but mis-targets one of them. It should
then allow them to recover from the issue by re-targeting one of the
two.

A few safeguards have been added to the feature so hopefully it
doesn't cause issues...

Fixes #1131