[IMP] runbot_merge: improve staging cancel message on PR reopen
Because reopening a PR updates its head in order to trigger a staging
cancel (in case the PR's batch is staged), the staging would
simultanously show "foo/bar#42 updated by XXX" and not have the PR at
all in its list of staged PRs, leading to a somewhat confusing even
though technically correct display on the dashboard.

If the head is updated and `closed` is being set to
`False` (indicating the PR is being reopened), set the reopening of
the PR as unstaging reason instead. It doesn't actually change
anything but should be less confusing to the reader who does not know
or remember about this subtlety e.g. me, myself, or I.

For the record,

- Updating the PR's head on reopening has been a thing since
  eb91e2371b0deff7806014b5abbd52b58bd14ef2 originally to revalidate on
  reopen (as otherwise statuses received while the PR was closed might
  not be taken in account).
- Unstaging on head update and thus on reopen as an implicit side
  effect was implemented in c8a06601a7c223cdfc9da23f2c799b4c660f274b.
- The behaviour gained feature status (via a test, and a documentation
  comment) in 62fbda52a8aa7b2ea6df0226c1b0bfbf27bb79cd.

Fixes #1130