[IMP] runbot_merge: prevent setting fw_policy / skipmerge from backend
The `fw=` command, especially `fw=skipmerge`, has a significant number
of side-effects which would be non-trivial and *extremeley* hacky to
set via setting the field via a backend form.

So just disable it. If it turns out to be a frequent need in the
future (as unlikely as that is), add a dedicated action and a button.