[IMP] runbot: base error month history on batch date
The `30 days history` widget shows the history based on the log_date
datetime. This datetime maybe innacurate when the related build took a
long time before starting.

With this commit, the datetime is for this history will be based on the
related batch create date.

The hourly recurrence graph is removed as it was not really useful.