[ADD] runbot_merge: self contained runsuite
This commit adds support for the pytest suite running mitmproxy and
dummy-central internally instead of *having* to run them externally.

Running them externally is still useful for debugging or tracing
purposes and might use less resources (for mitmproxy it's a bit hard
to quantify as it will likely use more memory but each instance will
have more work, for DC... it's not exactly severely loaded at the best
of time).