[IMP] runbot_merge: add per batch info to stagings endpoint
The `pull_requests` member provides an unstructured set of all the
pull requests in the staging.

This new member provides a list of batches (in-order, though not
entirely sure whether it's forward or backward), and the merged PRs
for each batch *including their merged heads* (aka which commit they
ended up in the repository / target branch as).