[ADD] *: support for mergiraf in stagings & forward ports
It probably won't change anyone's life, but some local testing and
looking at a few recent (conflicted) forward ports indicates that this
should reduce the number of conflicts by a non-insignificant amount.

Also enable it for stagings, I think structural non-conflicts are a
lot less of a concern there (though they do happen from time to time),
however because mergiraf is able to surface a few *logical* conflicts
it might be able to surface those more eagerly and avoid wasting time
running tests on something which is likely going to fail.

Fixes #1145