[FIX] runbot: fix threehash export date
The export date of the commit was broken since we are exporting
threehash instead of commit.

The main reason is that the initiall --mtime never worked, changing the
-xmC to -xC was the only needed change since git archive will set the
mtime of the file to the commit date by default.

To fix this issue, we need to add a --mtime on the git archive command

Note that since commit date are used, it is still possible to have a
date missmatch, since the first commit to export will set the date.

To fix this we may need to fix a more apropriate date, to define.

Note that issue already exist when restoring a database from another
branch (multi split, ...)

We should either use a filehash (cashed by date) when computing the
asset bundle hash, or export with the last file modification date if
possible