[IMP] runbot: allow to use the faketime tool
This commit allows to use the `faketime` command line tool to simulate a
build at fake date and/or time.

This can be done by passing a config_data like:

`
{
    'faketime': '2024-06-01 02:42 UTC'
}
`

Note that runbot will not suport advanced timestamp specification format
but only the simple DATESTRING format of the `/bin/date` utility.

Pay attention that this will only alter the datetime of the executed
Python code and its subprocesses, the postgresql server is not affected
by the faketime tool.

See:
- https://manpages.ubuntu.com/manpages/noble/man1/faketime.1.html
- https://manpages.ubuntu.com/manpages/questing/en/man1/date.1.html