[IMP] runbot: change disabled network strategy
When using the argument `network_disabled` from the python docker
interface, the `/etc/hosts` file is empty in the container and cannot
even be filled with the `extra_hosts` parameter. In this situtation, the
processes in the container are not able to resolve `localhost`.

With this commit, the `network_mode` is used instead. With this method,
the `/etc/hosts` file is properly filled.

Notes:
- The `/etc/hosts` file cannot be manipulated at the build time of the
  image as it's used by Docker itself during build process.
- The file cannot be filled at runtime as it needs to be the root user.