[FIX] runbot: properly strip spaces in test_tags
During a `run_install_odoo`step, spaces are removed from the custom
test-tags and the ones on the step itself.

If a parametric test-tag with spaces is given that way it leads to an
invalid test-tag and a build error.

With this commit, only the spaces at the beginning or the end of a
test-tag are cleared. A test is also added to validate that.

Finally, the `get_test_tags` utility method is fixed to properly find
the test-tags in the command line.