[IMP] forwardport: remove PR title prefix
It causes unnecessary heartache when using non-rebase merge
methods (as the prefix becomes part of a / the commit message), and
it's not really useful since the mergebot adds the relevant labels to
both forwardport and backport PRs, which is easier to search for and
classify. So it's really just noise.

Fixes #1153