[IMP] runbot_merge: flag staging disabled on main dashboard
Preparatory work for #1194: we're going to disable staging on the
branch if a staging ended up completely fucked, to give the admins
time to check on it and fix it up if necessary.

This means it'd probably be good if that state was visible easily,
which currently is not the case.