[FIX] runbot_merge: handle safety dance failure better
Before this, if the first phase of the safety dance succeeds but the
second fails we mark the staging as failed and try to re-stage the
PRs. However if the staging was *partially* (some branches succeeded)
then we might be re-staging PRs which are already in the branch, or
batches which are *partially* merged, none of which is good.

If we have second-staged any branch then instead of an FF failure we
now raise a different error, which can be handled separately by:

- failing the staging and all the batches which straddle the
  successful and failed repos
- marks the PRs which are only in the ported repositories as
  merged (because they are)
- keep the PRs only in the failed repos ready

Fixes #1194