[IMP] runbot_merge: only push actual commits to target
Previously, on staging success we would:

1. check that every repository's $target can be forward-ported from
   its current state (via tmp)
2. update every repository's $target

This would be done even if there was no PR staged for the
repository, triggering no-op fast-forwards. This second step is
*probably* unnecessary given (1) and increases increases both the time
it takes to "commit" a staging, and the risk that a github error will
wedge the staging.

This here commit changes this behaviour: we now base (1) on the heads,
and we only create commits when there are actual PRs staged on the
repository, thus when we fast-forward $target = $commit we only do so
for the repositories being staged for, not all repositories in the
project.

This makes the process slightly less safe in that a direct push in the
interval between the first and second steps of the safety dance can
lead to inconsistent / broken repository state, but it significantly
lowers the odds that github being wonky will trigger a push error on a
repo we don't have anything to push on anyway as happened in staging
138273.

Fixes #1194