[ADD] *: better non-webhook statuses
9c51f87aed36fc1f310e3854124d8767f65d871f previously added an API path
for validating stagings, however that had two limitations:

1. it only worked for stagings
2. it goes through the ORM for its RPC needs

The former is frustrating because it requires still hooking onto
statuses for PRs, which is not necessarily desirable, the second is an
issue because "some people" are looking to destroy it all, and access
is a pain anyway.

This new version changes tack completely by relying on internals'
internal role system for communication which takes care of
authentication. It also reveals a few limitations of some processes.

Aside from simplifying fw tests, it requires quite a bit of
conditional handling in the basic tests: going through the endpoints,
we can't set statuses on a commit which is not (yet) the head of a PR,
which means PRs end up `opened` rather than `validated`. One of the
tests which pretty much checks just that is simply xfailed in runbot
mode (for the rest these are mostly ancillary checks).

Also the "statuses changed" chatter messages are different so that
requires a conditional as well.

The effort from 9c51f87aed36fc1f310e3854124d8767f65d871f is kept for
now but I should probably remove it eventually, as it's never actually
been productized and there's basically no chances it will ever be.