[IMP] runbot_merge: oustanding PRs page
Fix handling of fw=skipmerge PRs: the "title" for a group / batch set
should be the source-less PR, even if it is not merged yet.

While at it move the hover info (creator and reviewer) of individual
PRs in the sequence to the `<li>` aka the PR link plus its state info
if any, as this requires less mousing precision for un-merged PRs.

Fixes #1204