[FIX] runbot_merge: fw=no on not the source PR
fw=no was kinda designed under the assumption that it would only be
set on the source PR, but apparently some devs have other idea.

Fix handling of forward-porting when triggered on a forward-port PR in
fw=no mode (forward porting should not happen), and fix branch
filtering when a PR is fw=no to take the existing genealogy instead of
just the current branch. That way if fw=no is (incorrectly) used as a
way to set a limit it doesn't break everything.

Fixes #1200