[ADD] runbot_merge: `nice` pr priority
New priority level is below `default` and can be used to de-prioritise
compared to base.

While at it fix the staging priority within a level:
`Batch.write_date` is modified way to often, including *while
staging* (probably), so it doesn't work right, or repeatedly. Add a
new `unblocked_at` field which stores the instant of the transition
from blocked to unblocked, and order by that. It should be
significantly more reliable and stable, and better represents "first
come first serve" (in this case first ready first staged).

Because `unblocked_at` should always be set if `blocked=False` the
fallback on sorting by `id asc` is probably unnecessary, but it should
also be mostly harmless.

Fixes #1211